snapenv: add UID environment variable
Some ENVs require the user ID, but the only way of getting it
is using a little shell script that exports the ENV after using
"id -u" to get the value. Other uses tricks, like taking advantage
of a similar path (like firefox, which sets PULSE_SERVER to
"$XDG_RUNTIME_DIR/../pulse/native".

This patch adds the SNAP_UID environment variable, which contains
the effective user id. With that, it's possible to define paths
containing it directly in the YAML file.