o/snapstate: pre-download and monitor snap for auto-refresh
Handle a pre-download task by downloading the snap (so it's in the cache
for the subsequent auto-refresh for that snap) and perform a soft check to
see if the snap is still running. If it is, notify the user to close the
snap and monitor for its closure. Once the snap closes, trigger an
auto-refresh change and clear the monitoring state. If the soft check
doesn't fail (the snap isn't running), the auto-refresh is triggered
directly.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>