New store-state tool
This is the new store tool inclusing a new test. Also it is included a
network-state tool

Still missing some parts of the test and a new test for the network
state tool. This can be added in a following pr including the migration
from the network-state helper to the new tool in all the tests