cmd: create the XDG_RUNTIME_DIR folder
Currently, the XDG_RUNTIME_DIR folder isn't created when a snap
is run, which requires in several cases to include a helper
script to create it if it doesn't exist.

This behavior violates the fd.o base directory spec
https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html
as commented here https://forum.snapcraft.io/t/rethinking-how-we-handle-xdg-runtime-dir/22223

This MR fixes this.