devicestate: fix missing ExtraSSLCerts support when getting serials
This commit fixes the issue that getting the serial fails in
confusing ways when a custom SSL certificate is needed to connect
to get the serial. The error message in this case is:
```
handlers_serial.go:659: Proxy store does not support custom serial vault; ignoring the proxy
```
which is not at all true.

The root cause for the bug is that the used http client does
set `ExtraSSLCerts`. However there are more issues with the
code, e.g. that `newEnoughProxy` does not return an error but
instead just debug lots it and does nothing less.

This commit is the minimal fix.