devicestate: change the MockDeviceService() to use https
There was a bug in snapd that prevented the serial handler code
to get a serial number when this required custom SSL certificates.

This case was not tested. This commit switches the MockDeviceService()
code to use https and to also write out the custom SSL certificate
into the `dirs.SnapdStoreSSLCertsDir` location.

By doing this we ensure that all interactions with the DeviceService
must also support custom SSL certificates.