devicestate: return an error from newEnoughProxy()
The previous code would not return errors from `newEnoughProxy()`
but instead only log them in a debug logger. This behavior is
problematic as a proxy error needs to be surfaced properly and
with this commit issues with incorrect SSL certificates for a
proxy are visible instead of hidden in the debug log.