configcore: allow to run core configuration on classic via env
This commit adds a new environment called
`SNAPD_BREAK_SYSTEM_BY_RUNNING_CORE_ONLY_CONFIG`
that allows to run core configuration options on classic.

The goal is to be able to test netplan integration on classic
systems. Nothing in netplan is core specific, we just did not
enable it on classic for conistency but it should be great
for netplan based integration testing inside the netplan repo.