snap: show log information from Wait tasks in the CLI (#12598)
* snap: show log information from Wait tasks in the CLI

This is a tiny patch to make the UX on classic+modes systems
more user friendly. Right now when running e.g.
`snap refresh pc-kernel` there is no indication that a system
reboot is required to continue, the only visible output is
that a re-refresh is attempted.

To fix this properly a bigger change is required and various
attempts in PR#12417 are all having problems. However as a
very minimal improvement this commit shows log messages from
tasks in `Wait` state without changing anything else.

This means that the UX is more like:
```
$ snap refresh pc-kernel
2023-02-24T11:2315Z INFO Task set to wait until a manual system restart allows to continue
Handling re-refresh of "pc-kernel" as needed
```
while not ideal UX it feels like a worthwhile improvement until
we find the time to fix this properly.

* snap: fix typo

Co-authored-by: Miguel Pires <miguelpires94@gmail.com>

* snap: add warning about pressing ctrl-c

* snap: tweak warning message when snaps are in "wait" state (thanks Samuele)

---------

Co-authored-by: Miguel Pires <miguelpires94@gmail.com>