testutil: add checker for nil typed and valued vars
Add an IsInterfaceNil checker that ensures both a variable's type and
value are nil. This is useful to prevent returned errors with non-nil
types from being compared to a nil value with a nil type (since the
compiler can't tell that the other variable is not nil typed), resulting
in the check failing.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>