Fix TestDeviceUnlockErr test in debian-sid and ubuntu 23.04
The unit test is failing in those systems with the following error:

fde_test.go:625:
c.Assert(err, ErrorMatches, `cannot run fde-device-unlock
"device-unlock":
    -----
    output-only-used-for-errors
    service result: exit-code
    -----`)
... error string = "cannot run fde-device-unlock \"device-unlock\": exit
status 1"
... regex string = "" +
...     "cannot run fde-device-unlock \"device-unlock\": \n" +
...     "-----\n" +
...     "output-only-used-for-errors\n" +
...     "service result: exit-code\n" +
...     "-----"

This is happens not 100% of the executions on those systems. It is
causes because the error message is not saved before the script exits
with 1. To deal with this race I added a short sleep in the test script
that is used to mock the fde-device-unlock to make sure the error
message is captured.