many: add new snapd-observe API interface via /run/snapd-ro.socket
This commit adds a new /run/snapd-ro.socket that and a new access
daemon API access checker called `observeAccess` that can be used
to make API functions available on the read-only socket.

In addition there is a new `snapd-observe` interface that allows
access to the new `/run/snapd-ro.socket` that will allow access
to this new socket.

Initially the snapd-ro socket will only grant access to the
`/v2/changes` API. Alternatively we could make everything under
the current `openAccess` available on the observe socket but
that needs a closer audit.