many: introduce min-size in gadget structures description
Introduce min-size so structures can specify it in the gadget. It will
be used in remodelling/gadget installation situation so we have wider
compatibility with older gadgets defining different sizes for the
partitions.

With this change we now cannot assume that we know the partitions
offsets in all cases and ranges will need to be taken into account.