Call prepare_memory_limit_override before the snapd state is saved
Avoid calling this function on every prepare because it is taking
several seconds and it is delaying the whole tests execution more than 1
hour.
Currently the tests are taking more than 2 hours instead of 70 minutes.

prepare_memory_limit_override is saving a configuration in
/etc/systemd/system/snapd.service.d/. This rerectory is being restored
during the prepare-each, so the change is to save the desired config
before the snapd state is saved, so we make sure the config is applied
for every test.