Add DBus interface for changes monitoring
This DBus interface allows an user-session program to know
when a delayed autorefresh has begun and ask for the current
status of the tasks, thus moving outside snapd all the code
to update the current refresh status.

This MR requires changes in snapd-desktop-integration.

This MR also adds some extra info in the internal query from the
user agent to the snapd daemon when asking data about a change.
The new data, added only when it is available, is the snap
name, instance name and snap revision of the snap being modified
by an specific task.