interfaces/builtin: custom-device kernel in subdir
Devices may be in subdirectories of `/dev/`, rather than directly in
`/dev/` itself.  This change allows matching the `kernel` attribute of
`udev-tagging` to the basename of a device path in `devices`, rather
than matching `/dev/${kernel}` to a device in `devices`.  Snapd maps the
value of the `kernel` key to the string used as the `KERNEL==` udev
rule.  To the best of my knowledge, this change makes custom-device udev
kernel validation match the behavior of the `KERNEL` udev rule.

Also adds an additional test which should fail on the `subsystem` key,
thus indicating that the `kernel` key was successfully validated.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>