Avoid reboot in snap-confine-tmp-mount test
The reboot is not needed in the test. Also the reboot was causing the
test auto-refresh-gating-from-snap to fail because the refresh was being
delayed about 5 minutes.

+ echo 'Trigger auto-refresh of test-snapd-refresh-control-provider but
hold it via test-snapd-refresh-control'\''s hook'
Trigger auto-refresh of test-snapd-refresh-control-provider but hold it
via test-snapd-refresh-control's hook
+ /home/gopath/src/github.com/snapcore/snapd/tests/lib/tools/snapd-state
force-autorefresh
+ systemctl start snapd.service snapd.socket
++
/home/gopath/src/github.com/snapcore/snapd/tests/lib/tools/snapd-state
wait-for-autorefresh 1
+ LAST_REFRESH_CHANGE_ID='snapd-state: expected a new auto-refresh
change with id greater than 1, but it didn'\''t happen'