interfaces/builtin: improved mount option handling
This commit changes fs-specific option handling so that the option
validation fails if a mount option is _not_ one of:
1) a kernel mount option,
2) a userspace mount option, or
3) a fs-specific mount option which is compatible with every specified
   filesystem for the given mount configuration.

Previously, mount options were allowed if they were found on the option
list for any of the specified filesystems.  This was not correct, as
mount options should only be allowed if they are allowed for all the
possible filesystems for a given mount specification.

Additionally, this commit adjusts the corresponding unit tests to revert
to their original form those which existed before the userspace mount
options changes were added, and to explicitly add new tests which check
userspace mount options (`nofail` in particular) and filesystem-specific
mount options according to the rules described above.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>