interfaces/builtin: removed `nofail` from kernel mount options
The `nofail` mount option is a userspace option which is not passed to
the kernel, and thus should not be included in the apparmor mount rules.

Since previous commits related to userspace mount options introduced
mount option filtering to include only kernel mount options in the
apparmor mount rules (after all mount options are validated), it is
important that `nofail` not be included in the kernel mount option list.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>