interfaces/builtin: differentiate mount options with{,out} optargs
By splitting after the `=` in mount options such as `uid=1000`, we can
differentiate an allowed mount option which requires an optarg, such as
`uid=1000` for functionfs, from `uid`, which is not allowed.  For
options with optional optargs, such as btrfs `compress`, by including
both `compress` and `compress=` in the list of allowed filesystem mount
options.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>