prompting/storage: split db into file, dir, and subdir decisions
This adds handling for the "allow-directory" and "allow-subdirectories"
options, and treats calls to `Set()` without either as allowing access
to the given file (or directory) only.

When attempting to match a path to an existing decision, first search
for a direct match in the Allow, AllowWithDir, or AllowWithSubdirs maps.
Then, search for a direct match for the parent of the path in the
AllowWithDir and AllowWithSubdirs maps. Lastly, search for a match to
any ancestor of the path in the AllowWithSubdirs map.

The `Set()` function should never be able to store the same path in
multiple maps.  However, if this does occur, then if a search for a path
results in matches within multiple maps at the same time (that is, both
match exact path, both match parent, or both match same ancestor), an
error will be thrown.

Notably, this change now treats file-only permissions on directories as
only allowing access to the directory file itself, not any of the files
contained in the directory.  My interpretation is that this permits
listing the contents of the directory but not accessing any of the
listed files.

This is a first draft of this feature, and there remains work to be done
to fix the existing unit tests as well as dramatically expand them to
cover the many scenarios which can occur when searching for or
overwriting previous decisions.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>