prompting/storage: fixed regression due to cleaning path
The `filepath.Clean()` function strips trailing `/` characters, thus
eliminating the extra indication to treat the given path as a directory,
rather than a file, and thus not call `filepath.Dir(path)` when handling
"allow-directory" and related information.

Cleaning is still important, however, so this change moves it to after
the `filepath.Dir()` would occurs.

Additionally, in `findPathInLabelDB()`, the extra call to
`filepath.Dir()` if the path ends in `/` is removed, since
`filepath.Clean()` handles this.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>