prompting/storage: simplified path map match handling
Removed the `jsonAllow{,With{Dir,Subdirs}}` const variables which were
returned by `findPathInLabelDB()` and used by `Set()` to check in which
map a match was found.  In its place, `findPathInLabelDB()` returns a
pointer to the map itself, thus reducing both the number of string
consts which must be kept up to date as well as the number of string
comparisons, which have greater room for coder error than a type checked
variable comparison.

As a nice side effect, exact matches can now be removed directly from
the map in which they were found, rather than attempting to remove the
path from all maps (no-op if the path is not in the map).

The only downside is in the case where a path matches multiple maps
simultaneously (that is, either directly or at the same ancestor).
In this case previously, `findPathInLabelDB()` returned the json names
of each of the maps which matched, separated by commas, which could in
the future be used as part of error handling or messaging.  However, it
should be impossible for `Set()` to ever assign the same path in
multiple maps, so this should only occur if there is external
interference with the json file.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>