interfaces/builtin: custom-device kernel in subdir (#12734)
* interfaces/builtin: custom-device kernel in subdir

Devices may be in subdirectories of `/dev/`, rather than directly in
`/dev/` itself.  This change allows matching the `kernel` attribute of
`udev-tagging` to the basename of a device path in `devices`, rather
than matching `/dev/${kernel}` to a device in `devices`.  Snapd maps the
value of the `kernel` key to the string used as the `KERNEL==` udev
rule.  To the best of my knowledge, this change makes custom-device udev
kernel validation match the behavior of the `KERNEL` udev rule.

Also adds an additional test which should fail on the `subsystem` key,
thus indicating that the `kernel` key was successfully validated.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* interfaces/builtin: use filepath.Base instead of custom code

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* interfaces/builtin: fixed udev kernel device matching

This change allows udev-tagging rules to have `kernel: deviceName` values
which match the end of device paths, rather than necessarily matching
either `"/dev/" + deviceName` or `deviceName` matching the basename of
the device.

Before, a basic udev `KERNEL==` rule was generated for each device and
then overwritten by the `kernel: deviceName` rules in the `udev-tagging`
section which matched `"/dev/" + deviceName` to the device path.  Now,
the `kernel: deviceName` rules are converted to udev rules first, and
then any remaining devices which have not been matched by a rule are
given the basic `KERNEL==` udev rule.

If the number of devices are matched by the same `deviceName` from a
`kernel: deviceName` specification is not 1, then an error is thrown.
This is similar to previous handling of `deviceName` not matching any
devices, but now also includes checking that `deviceName` does not match
more than one device, as this is presumably not desired.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* interfaces/builtin: custom-device stricter udev kernel rules

This change restricts the `udev-tagging` `kernel` attribute matching
rules so that `kernel: deviceName` only matches a `devicePath` from the
device list of `"/dev/"+deviceName == devicePath` or `deviceName` is
equal to the basename of `devicePath`.

Previously, `deviceName` matched `devicePath` if `deviceName` was the
string suffix of `devicePath` (not suffix in the path sense).  This
allowed, for example, `bar/name` to match `/dev/foo/bar/name`, where
now it would only match `/dev/bar/name`.

It is still the case that a `deviceName` might match more than one
device, in which case an error will be thrown.  For example,
`kernel: name` would match every device in
`devices: [/dev/name, /dev/foo/name, /dev/bar/name]`, which would throw
an error, since only one match is allowed for a given kernel
`deviceName`.

Note: if `deviceName` contains a subdir (e.g. `foo/bar`), then it must
match a complete path relative to `/dev/`, since it will never match the
basename of a device.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* interfaces: make UDevConnectedPlug slightly more readable

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>
Co-authored-by: Michael Vogt <mvo@ubuntu.com>