o/devicestate: create pending users after receiving a serial
Currently, if you attempt to create a user using a serial-bound
assertion before the serial is available, the request will simply
fail. This commit makes it so that attempts to create a system-user with a
serial-bound assertion fail, but a flag is set within the device
manager's state. Once the serial assertion is acquired, any serial-bound
users will be created by the device manager's Ensure() function.

Rather than relying on Ensure(), we could probably place this
functionality inside the FinishRegistration() function, since it should
only be called once; however, I opted to keep FinishRegistration()
slim and readable.

Signed-off-by: Joseph Brock <joseph.brock@canonical.com>