prompting/storage: added per-permission decision handling
This change introduces handling to allow or deny particular file
permissions on filepaths.  This is done by adding another map layer
between label and paths: for each label, there is a map which maps from
permission (`string`) to `*permissionDB`, where the original `Allow`,
`AllowWithDir`, and `AllowWithSubdirs` maps still reside.

Thus, the allow maps for different permissions are fully independent.
For example, one could set `/home/test/Documents` to allow read access
with subdirectories, but write access only with directory (that is,
entries directly in `/home/test/Documents/`).

At the moment, the requests passed into `Get()` and `Set()` contain a
`Permission` field which is populated from type
`apparmor.FilePermission`, which is (for now) a bitflag.  These bitflags
are converted into a `[]string` where each element is the string
corresponding to the permission bit as defined in
`prompting/apparmor/permission.go`.

Since the `Get()` function only has a `notifier.Request` as input, the
permission in question can be extracted from it according to the method
described above.  For `Set()`, however, there is the additional
`extras["{allow,deny}-extra-permissions"]` parameter to handle as well.
For now, these permissions are expected to be comma-separated strings
which again correspond to the permission bits as defined in
`prompting/apparmor/permission.go`, and they are appended to those
permissions extracted from the `Permission` field of the
`notifier.Request`, and thus treated no differently.

It is assumed that only one permission bit should be set in the
`Permission` field of the `notifier.Request`, but handling has been
added to `Get()` in case multiple bits are set -- in this case, every
permission must be allowed, else the `allow` field is returned as
`false`.  `Set()` needs no additional special handling, as it is already
assumed that there may be multiple permissions defined between the
`Permissions` and `extras["{allow,deny}-extra-permissions"]`.

It is, however, assumed that the same allow/deny decision must be set to
all the permissions passed into `Set()` (via the `req.Permissions` field
or the `extras` parameter) at a time.  To allow some permissions for a
path and deny others, `Set()` may be called multiple times with
different values for the `allow` parameter.

Alternatively, a future approach might be to respect
`extras["allow-extra-permissions"]` and
`extras["deny-extra-permissions"]` independently of the value of the
`allow` parameter.  However, at the moment, the apparmor prompting spec
indicates that the `"allow-*"` and `"deny-*"` keys may only be used when
`allow = true` and `allow = false`, respectively.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>