github: remove use of `covertool merge` to fix codecov reporting (#12837)
The codecov reporting was broken for a while and after some
slightly painful debugging in PR#12834 it turns out that
the use of `covertool merge` seems to break codecov. Using
this is not needed as codecov can merge reports natively.

This commit removes the use of `covertool merge` and instead
just send the various generated coverage reports to codecov.