interfaces/builtin: alphabetized kernel mount options
This change alphabetizes the supported kernel mount options, as well as
moves `zfs` after `xfs` in the list of default filesystems.

Filesystem-specific mount options are not alphabetized, and are instead
left in the order they are defined in their documentation or source
code in order to ease visual verification of their correctness.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>