packagekit: allow dbus activation from snapped apps
Currently, a snapped application with the "packagekit-control"
plug connected (like snap-store/ubuntu-software), can access the
packagekit daemon through DBus without problem... as long as the
daemon is already running. If it isn't running, DBus activation
won't work for a snapped application.

A way of checking this is by launching a shell inside the snap-
store snap

    snap run --shell snap-store

and do a DBus call, like

    dbus-send --system --print-reply --dest=org.freedesktop.PackageKit /org/freedesktop/PackageKit org.freedesktop.DBus.Properties.GetAll string:"org.freedesktop.PackageKit"

If packagekitd is running, this call will succeed, but if it's
not running, it will fail.

This problem was detected by the people of the new snap-store.
Currently they use a trick to ensure that packagekitd is
launched, that consists on:

    dbus-send --system --print-reply --dest=org.freedesktop.DBus /org/freedesktop/DBus org.freedesktop.DBus.StartServiceByName string:org.freedesktop.PackageKit uint32:0

This PR fixes this, although it needs security review.